requirements.txt
flask==3.0.0
alpaca-py==0.26.0
pandas==2.2.0
numpy==1.26.0
requests==2.31.0
APScheduler==3.10.4
gunicorn==21.2.0
